backend/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dj-database-url==2.2.0
Django==4.2.16
django-cors-headers==4.4.0
django-spa==0.3.6
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.18.1
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
numpy==2.0.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.13.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
whitenoise==5.3.0
zipp==3.20.2
